fix: pass all loaders in `LoadConfig`
