+++ gonk_test.go
@@ -1,5 +1,11 @@
 package gonk
 
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
 type IntermediateA struct {
 	FieldE string `config:"fieldE"`
 }
@@ -15,4 +21,35 @@ type RootType struct {
 	FieldC string          `config:"fieldC,optional"`
 	FieldD IntermediateA   `config:"fieldD"`
 	FieldF []IntermediateB `config:"fieldF,optional"`
+}
+
+func TestMultiLoader(t *testing.T) {
+	assert := assert.New(t)
+	out := new(RootType)
+	expected := RootType{
+		FieldA: "hello",
+		FieldB: 10,
+		FieldD: IntermediateA{
+			FieldE: "world",
+		},
+		FieldF: []IntermediateB{
+			{FieldG: "foo", FieldH: "bar"},
+			{FieldG: "baz"},
+		},
+	}
+
+	t.Setenv("CONFIG_FIELDB", "10")
+	t.Setenv("CONFIG_FIELDD_FIELDE", "world")
+
+	mapLoader := MapLoader(map[string]any{
+		"fieldA": "hello",
+		"fieldD": map[string]any{},
+		"fieldF": []any{
+			map[string]any{"fieldG": "foo", "fieldH": "bar"},
+			map[string]any{"fieldG": "baz"},
+		},
+	})
+	envLoader := EnvLoader("config")
+	assert.NoError(LoadConfig(out, mapLoader, envLoader))
+	assert.Equal(expected, *out)
 }
